(u_int) is the wrong type here.  Use unsigned long instead, even though that's only less wrong...
# This gets the kernel building again to the point it was at before
# the last IFC for the OCTEON1 kernel config.


Former-commit-id: 10b99b7b8f296ffccf0da869dab6fee99ee71590

+++ sys/mips/include/param.h
@@ -197,7 +197,7 @@ name|_ALIGN
 parameter_list|(
 name|p
 parameter_list|)
-value|(((u_int)(p) + _ALIGNBYTES)&~ _ALIGNBYTES)
+value|(((unsigned long)(p) + _ALIGNBYTES)&~ _ALIGNBYTES)
 end_define
 
 begin_define
@@ -230,7 +230,7 @@ name|p
 parameter_list|,
 name|t
 parameter_list|)
-value|((((unsigned)(p))& (sizeof (t) - 1)) == 0)
+value|((((unsigned long)(p))& (sizeof (t) - 1)) == 0)
 end_define
 
 begin_comment
